+++ SignNow.Net/Interfaces/IUserService.cs
@@ -14,10 +14,10 @@ namespace SignNow.Net.Interfaces
         /// <summary>
         /// Creates an account for a user
         /// </summary>
-        /// <param name="user">User personal data (firstname, lastname, email, password)</param>
+        /// <param name="createUser">User personal data (firstname, lastname, email, password)</param>
         /// <param name="cancellation">Propagates notification that operations should be canceled</param>
         /// <returns></returns>
-        Task<UserCreateResponse> CreateUserAsync(UserRequest user, CancellationToken cancellation = default);
+        Task<UserCreateResponse> CreateUserAsync(CreateUserOptions createUser, CancellationToken cancellation = default);
 
         /// <summary>
         /// Retrieve current user`s resource
@@ -25,5 +25,13 @@ namespace SignNow.Net.Interfaces
         /// <param name="cancellationToken">Propagates notification that operations should be canceled</param>
         /// <returns></returns>
         Task<User> GetCurrentUserAsync(CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Updates user information i.e. first name, last name
+        /// </summary>
+        /// <param name="updateUser">User personal data (firstname, lastname, password)</param>
+        /// <param name="cancellationToken">Propagates notification that operations should be canceled</param>
+        /// <returns></returns>
+        Task<UserUpdateResponse> UpdateUserAsync(UpdateUserOptions updateUser, CancellationToken cancellationToken = default);
     }
 }